[balance] fixed bug for LARGE matrices.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@1955 de4796e2-0b1e-0410-91b4-86bc23549317

+++ linalg/balance.c
@@ -1850,7 +1850,7 @@ void weight_sort_hslice(sink datasink, fileset fs, unsigned int ii)
      */
 
     size_t * nbytes = malloc(nrs * sizeof(size_t));
-    uint32_t tot = 0;
+    size_t tot = 0;
     for(i = 0 ; i < nrs ; i++) {
         nbytes[shuffle[i]]=line_bytes[row_slices[ii].r[i]];
     }
